Made Polygon be drawn correctly.

+++ src/test/java/PointTest.java
@@ -5,8 +5,8 @@ public class PointTest {
 
     @Test
     public void testString() {
-        double xCoordinate = 10.0;
-        double yCoordinate = 20.0;
+        int xCoordinate = 10;
+        int yCoordinate = 20;
         Point point = new Point(xCoordinate, yCoordinate);
         String expectedOutput = "(" + xCoordinate + ", " + yCoordinate + ")";
         assertEquals(expectedOutput, point.toString());
